Fix data call issue for single pdn case.
When a higher priority apncontext is scheduled for retry, default
data cannot be setup if only single pdn is allowed. If the higher
priority apncontext is later cleaned up before the retry alarm expires,
default data is never re established.
To fix this, try setup data on connectable apns after an apncontext
is cleaned up.

Change-Id: I80e97f6d862ef27e0237db332280e7296d9eebe9
CRs-Fixed: 715866, 857421, 818834