MSIM: Fix to handle timeout related finish response
When set radio cap request timeout happens, ProxyController sends FINISH
with failure status to cleanup ongoing transcation, while sending
FINISH reuqests it generates new session id, but the generated new
session id is not updated in global variable and because of that the
FINISH response is ignored.

To fix this update the generated new session id in global variable.

Change-Id: I4ec7aa106f65a763809bd3d9392b56882216f132